Parameters units are converted in more accurate place
