+++ packages/medical-domain/src/domains/repositories/medical/datum/device-parameter-change.service.ts
@@ -30,6 +30,7 @@ import { DatumProcessor } from '../../../models/medical/datum.model'
 import BaseDatumService from './basics/base-datum.service'
 import MedicalDataOptions from '../../../models/medical/medical-data-options.model'
 import Unit from '../../../models/medical/datum/enums/unit.enum'
+import { getConvertedParamUnitAndValue } from '../../../utils/unit.util'
 
 /**
  * Used to regroup device parameters in one tooltip, when the changes are too close.
@@ -40,6 +41,8 @@ const DEVICE_PARAMS_OFFSET = 30 * 60 * 1000
 
 const normalize = (rawData: Record<string, unknown>, opts: MedicalDataOptions): DeviceParameterChange => {
   const base = BaseDatumService.normalize(rawData, opts)
+  const { unit, value } = getConvertedParamUnitAndValue(rawData.units as Unit, rawData.value as string, opts.bgUnits)
+  const { value: previousValue } = getConvertedParamUnitAndValue(rawData.units as Unit, rawData.previousValue as string, opts.bgUnits)
   const deviceParameterChange: DeviceParameterChange = {
     ...base,
     type: 'deviceEvent',
@@ -52,9 +55,9 @@ const normalize = (rawData: Record<string, unknown>, opts: MedicalDataOptions): 
         uploadId: rawData.uploadId as string,
         name: rawData.name as string,
         level: rawData.level as string,
-        units: rawData.units as Unit,
-        value: rawData.value as string,
-        previousValue: rawData.previousValue as string,
+        units: unit as Unit,
+        value,
+        previousValue,
         lastUpdateDate: rawData.lastUpdateDate as string
       }
     ]
